+++ src/frontends/term/include/terminal_state.hpp
@@ -12,10 +12,10 @@ enum InputMode {
 
 typedef struct {
   int cy, cx;
-  int info_col_size = 0;
+  uint16_t info_col_size = 0;
   uint64_t line_offset = 0, frame_num = 0;
   uint64_t current_action_multiplier = 1;
-  int nrows, ncols;
+  uint16_t nrows, ncols;
   InputMode input_mode;
   std::string raw_input;
   bool reached_eof = false;
